Added corrections on story response code & types (see project website)

+++ payment/src/main/java/ca/ulaval/glo4002/payment/domain/payment/Payment.java
@@ -1,12 +1,10 @@
 package ca.ulaval.glo4002.payment.domain.payment;
 
-import java.math.BigDecimal;
-
 public class Payment {
 
-  private BigDecimal amount;
+  private float amount;
 
-  public BigDecimal getAmount() {
+  public float getAmount() {
     return amount;
   }
